Fix Rails CVE-2012-2661-like issues in Squeel.
Additionally, I've done some cleanup of areas where I don't
think I need to be calling to_sym, to reduce the risk of a
potential memory leak or DoS via malicious params that get
passed without any checks in user code. While I still
believe that protection against this sort of thing should
be the responsibility of the user, that doesn't mean that
I won't put a safety on the gun to reduce the risk of
someone shooting themselves in the foot.

While all specs are passing with these changes, I'm hoping
to get some folks to try this out in their apps before I
release a new version of the gem.