Remove a passthrough for String.
Should be unnecessary to remove this, but this fixes #162 and I honestly
don't remember why I added the passthrough to begin with. ARel handles
quoting/escaping on String so passing it through from Squeel shouldn't
have an effect. Since no specs fail when I remove the line, I'm gonna
give myself a good slap on the wrist for writing code that isn't
justified by specs, remove the line, and close #162 in good conscience.