The where_unscoping test case does apply to Rails 4.2
Perhaps the original author thought it didn't because there was no
`where_unscoping` override defined for 4.2 but it inherits from 4.1.

It doesn't hurt to do the test anyway, except that the unscope call
had to be changed to `unscope!` in order for the SQL check to pass. It
should have been `unscope!` all along and it only worked because of
rails/rails#20722.