Fix where_unscoping on 4.0, 4.1, 4.2
Believe it or not, basic `unscope` functionality was broken. I don't
know how but With Squeel in play, `where_values` somehow becomes a
nested array. Flattening it first fixes the issue.

I have adjusted the 4.1 override to use `super` so that less code is
duplicated. This isn't possible for 4.0. I have added a new override
for 4.2 because `where_values` should not be mutated in that version.