Marks that gem is no longer actively maintained, invites new maintainer or migration option.
Ernie mentioned in mid 2016 that he was happy to put a reference to baby_squeel in the project if the maintainers were no longer willing to maintain this project https://github.com/activerecord-hackery/squeel/issues/412. It's been three years and no new commits. I think this will be helpful to invite more maintainers as well as give information for those upgrading past 4.